Remove will-change property
Including `will-change` on `.glide__slides` introduces a rendering issue
in Safari when a slide contains a video and the user tries to watch said
video in full screen. Removing this one property (or defaulting it to
`auto`) resolves the issue.

See https://codepen.io/robmaurizi-the-vuer/full/rNJpPON for an example
of the offending behavior.