Add a11y attr option for hidden/visible slides
This will enable a default option that will set a visible class to all
elements in the current view and set aria-hidden on slides that are not
in view. This will also apply a negative tab index for all focusable
elements for slides that are not currently in the view.